test: add additional LSP20 tests
